pyproject.toml
[project]
name = "score-management"
version = "0.1.0"
description = "Manage Score Of Student"
dependencies = [
    "python-dotenv",
    "psycopg2-binary"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/utils", "src/models", "src/services", "src/views"]